Remove sp file notion from compiler Signed-off-by: Avi Moradi <avi.moradi@hp.com>
