+++ score-lang-api/src/main/java/org/openscore/lang/api/Slang.java
@@ -53,8 +53,8 @@ public interface Slang {
 	public Long compileAndRun(SlangSource source, Set<SlangSource> dependencies, Map<String, ? extends Serializable> runInputs, Map<String, ? extends Serializable> systemProperties);
 
 	/**
-	 * Load system property sources written in slang and map them to fully qualified names
-	 * @param sources the slang sources containing the system properties
+	 * Load system property sources written in yaml and map them to fully qualified names
+	 * @param sources the {@link org.openscore.lang.compiler.SlangSource}s containing the system properties
 	 * @return map containing all of the system properties with fully qualified keys
 	 */
 	public Map<String, ? extends Serializable> loadSystemProperties(SlangSource... sources);
