+++ score-lang-api/src/main/java/org/openscore/lang/api/SlangImpl.java
@@ -21,6 +21,7 @@ import org.openscore.lang.entities.CompilationArtifact;
 import org.openscore.lang.entities.ScoreLangConstants;
 import org.openscore.lang.runtime.env.RunEnvironment;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.yaml.snakeyaml.Yaml;
 
 import java.io.Serializable;
 import java.util.HashMap;
@@ -56,6 +57,8 @@ public class SlangImpl implements Slang {
     private Score score;
     @Autowired
     private EventBus eventBus;
+    @Autowired
+    private Yaml yaml;
 
     @Override
     public CompilationArtifact compile(SlangSource source, Set<SlangSource> dependencies) {
@@ -93,9 +96,29 @@ public class SlangImpl implements Slang {
 		return run(compilationArtifact, runInputs, systemProperties);
 	}
 
+//    private Map<String, ? extends Serializable> system_properties;
+//    public Map<String, ? extends Serializable> getSystemProperties() {
+//        return system_properties;
+//    }
+
 	@Override
 	public Map<String, ? extends Serializable> loadSystemProperties(SlangSource... sources) {
-		return compiler.loadSystemProperties(sources);
+		Validate.notNull(sources, "You must supply a source to load");
+		Map<String, Serializable> result = new HashMap<>();
+		for(SlangSource source : sources) {
+			Map<String, ?> parsed = (Map<String, ?>)yaml.load(source.getSource());
+			String namespace = (String)parsed.get("namespace");
+			Validate.notNull(namespace, "No namespace specified");
+			Map<String, ? extends Serializable> systemProperties = (Map<String, ? extends Serializable>)parsed.get("system_properties");
+			Validate.notNull(systemProperties, "No system properties specified");
+//			ParsedSlang parsedSlang = yamlParser.parse(source);
+//			String namespace = parsedSlang.getNamespace();
+//			Map<String, ? extends Serializable> systemProperties = parsedSlang.getSystemProperties();
+			for(Map.Entry<String, ? extends Serializable> entry : systemProperties.entrySet()) {
+				result.put(namespace + "." + entry.getKey(), entry.getValue());
+			}
+		}
+		return result;
 	}
 
     @Override
