+++ score-lang-compiler/src/main/java/org/openscore/lang/compiler/SlangCompiler.java
@@ -8,12 +8,10 @@
  */
 package org.openscore.lang.compiler;
 
+import java.util.Set;
+
 import org.openscore.lang.compiler.modeller.model.Executable;
 import org.openscore.lang.entities.CompilationArtifact;
-
-import java.io.Serializable;
-import java.util.Map;
-import java.util.Set;
 
 public interface SlangCompiler {
 
@@ -29,15 +27,7 @@ public interface SlangCompiler {
      * Pre-compile a Slang source into an {@link org.openscore.lang.compiler.modeller.model.Executable}
      * @param source the {@link org.openscore.lang.compiler.SlangSource}
      * @return an {@link org.openscore.lang.compiler.modeller.model.Executable} object, containing either a flow or operation in the file.
-     *          Returns null if the source contains Slang system properties
      */
     Executable preCompile(SlangSource source);
-
-    /**
-     * Load system property sources written in slang and map them to fully qualified names
-     * @param sources the {@link org.openscore.lang.compiler.SlangSource}s containing the system properties
-     * @return map containing all of the system properties with fully qualified keys
-     */
-    Map<String, ? extends Serializable> loadSystemProperties(SlangSource... sources);
 
 }
